correções pós feedback

+++ src/business/DeleteBikeByIdBusiness.ts
@@ -1,14 +1,15 @@
 import { DeleteBikeByIdDatabase } from "../data/DeleteBikeByIdDatabase";
 
-
 export class DeleteBikeByIdBusiness {
+  constructor(private deleteBikeByIdDatabase: DeleteBikeByIdDatabase) {}
 
-  constructor(
-    private deleteBikeByIdDatabase:DeleteBikeByIdDatabase
-  ){}
-    
-  deleteBike = async (id:string): Promise<void> => {
+  deleteBike = async (id: string): Promise<void> => {
+
     try {
+      const registeredBike = await this.deleteBikeByIdDatabase.findById(id);
+      if (registeredBike.length === 0) {
+        throw new Error("Bike não cadastrada.");
+      }
 
       await this.deleteBikeByIdDatabase.deleteBike(id);
